requirements.txt
Flask==2.3.3
mysql-connector-python==9.0.0
